requirements.txt
datasets==4.4.1
hydra-core==1.3.2
peft==0.18.0
torch==2.9.1
